requirements.txt
supabase
pandas
python-dotenv
requests